refactor: specをelectricity_charge_simulatorsに合わせて命名変更
