+++ serverside_challenge_1/challenges/jibiking/spec/support/electricity_charge_simulators_helper.rb
